examples: Cleanup Python examples
Cleanup Python examples
