+++ examples/python/led.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
-# Author: Thomas Ingleby <thomas.c.ingleby@intel.com>
-# Copyright (c) 2014 Intel Corporation.
+# Author: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
+# Copyright (c) 2018 Linaro Ltd.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -21,10 +21,35 @@
 # LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+#
+# Example Usage: Reads maximum brightness value for user1 led and turns it
+#                on/off depending on current state. Then sets led trigger
+#                to heartbeat
 
 import mraa
+import time
 
-print (mraa.getVersion())
-x = mraa.Gpio(13)
-x.dir(mraa.DIR_OUT)
-x.write(1)
+# initialise user1 led
+led_1 = mraa.Led("user1")
+
+# read maximum brightness
+val = led_1.readMaxBrightness()
+
+print("maximum brightness value for user1 led is: %d" % val);
+
+# turn led on/off depending on read max_brightness value
+if (val >= 1):
+    val = 0
+# never reached mostly
+else:
+    val = 1
+
+# set LED brightness
+led_1.setBrightness(val)
+
+# sleep for 5 seconds
+time.sleep(5)
+
+led_1.trigger("heartbeat")
+
+print("led trigger set to: heartbeat")
