+++ examples/python/uart.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
-# Author: Brendan Le Foll <brendan.le.foll@intel.com>
-# Copyright (c) 2014 Intel Corporation.
+# Author: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
+# Copyright (c) 2018 Linaro Ltd.
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -20,34 +20,21 @@
 # NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
 # LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE
+# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+#
+# Example Usage: Sends data through UART
 
-import mraa as m
+import mraa
+import time
+import sys
 
-# this example will show the 'advanced' i2c functionality from python i2c
-# read/write
+# serial port
+port = "/dev/ttyACM0"
 
-x = m.I2c(0)
-x.address(0x77)
+data = 'Hello Mraa!'
 
-# initialise device
-if x.readReg(0xd0) != 0x55:
-  print("error")
+# initialise UART
+uart = mraa.Uart(port)
 
-# we want to read temperature so write 0x2e into control reg
-x.writeReg(0xf4, 0x2e)
-
-# read a 16bit reg, obviously it's uncalibrated so mostly a useless value :)
-print(str(x.readWordReg(0xf6)))
-
-# and we can do the same thing with the read()/write() calls if we wished
-# thought I'd really not recommend it!
-
-x.write(bytearray(b'0xf40x2e'))
-
-x.writeByte(0xf6)
-d = x.read(2)
-
-# WARNING: python 3.2+ call
-print(str(d))
-print(int.from_bytes(d, byteorder='little'))
+# send data through UART
+uart.write(bytearray(data, 'utf-8'))
